Add PhoneGap 9.0.0 with Android SDK 30
Change the versioning scheme to include the SDK version instead of the
Cordova Android platform version. In fact, the platform is just a
detail that follows the fact of using a given Android SDK.

Furthermore, a user of the image may use phonegap to install another
platform version. The version used when building the image only affects
the cached npm packages.

Update Gradle to 6.5. It does not make sense to have an older Gradle
compared to the webratio/phonegap:7.1.1_android-30 image.