Update gitblit image for the latest jmoger/gitblit version
The latest is now at 1.8.0. It should not move again, since
jmoger/gitblit is migrated to gitblit/gitblit.
Keep using jmoger/gitblit to avoid the issues of the new version around
the port number change from 80 to 8080.

Align the name of the default properties file.

Remove the CMD command and just inherit the startup command from the
base image. The old CMD command does not work on newer jmoger/gitblit
anyway.