Add OpenJDK 8 image
The old webratio/java:8 image cannot be built anymore because it is
based on Oracle JDK. Oracle JDK 8 is no longer available.

Switch to OpenJDK to work around the issue.
Reflect the fact that we are using OpenJDK in the tag name, so we can
use other Java distributions if the need arises.